+++ src/App.tsx
@@ -5,6 +5,8 @@ import { Footer } from './components/Footer';
 import './styles/blocks/container.scss';
 
 export const App = () => {
+
+
   return (
     <div className="app-wrapper">
       <Header />
