+++ src/components/Root/Root.tsx
@@ -4,7 +4,7 @@ import { HomePage } from '../../pages/HomePage/HomePage';
 import { CatalogPage } from '../../pages/CatalogPage/CatalogPage';
 import { FavouritesPage } from '../../pages/FavouritesPage/FavouritesPage';
 import { ItemCardPage } from '../../pages/ItemCardPage/ItemCardPage';
-import { CartPage } from '../../pages/CartPage/CartPage';
+// import { CartPage } from '../../pages/CartPage/CartPage';
 
 export const Root = () => (
   <Routes>
@@ -20,7 +20,7 @@ export const Root = () => (
       <Route path="item">
         <Route path=":itemId" element={<ItemCardPage />} />
       </Route>
-      <Route path="cart" element={<CartPage />} />
+      {/* <Route path="cart" element={<CartPage />} /> */}
     </Route>
   </Routes>
 );
