+++ src/components/ShopCategory/ShopCategory.tsx
@@ -2,9 +2,11 @@ import React from 'react';
 import './ShopCategory.scss';
 import '../../styles/utils/text-styles.scss';
 import { Category } from '../../enums/Category';
+import { useAppSelector } from '../../app/hooks';
 
 export const ShopCategory: React.FC = () => {
   const categories = Object.values(Category);
+  const { products } = useAppSelector(state => state.productsReducer);
 
   return (
     <section className="shop" id="shop">
@@ -19,34 +21,11 @@ export const ShopCategory: React.FC = () => {
             <h4 className="product__title h4">
               {category === Category.Phones ? 'Mobile Phones' : category}
             </h4>
-            <span className="product__counter">95 models</span>
+            <span className="product__counter">
+              {`${products.filter(product => product.category === category).length} models`}
+            </span>
           </article>
         ))}
-
-        {/* <article className="product shop__product">
-          <div className="product__photo-container">
-            <span className="product__photo product__photo--phones" />
-          </div>
-
-          <h4 className="product__title h4">Mobile phones</h4>
-          <span className="product__counter">95 models</span>
-        </article>
-
-        <article className="product shop__product">
-          <div className="product__photo-container">
-            <span className="product__photo product__photo--tablets" />
-          </div>
-          <h4 className="product__title h4">Tablets</h4>
-          <span className="product__counter">24 models</span>
-        </article>
-
-        <article className="product shop__product">
-          <div className="product__photo-container">
-            <span className="product__photo product__photo--accessories" />
-          </div>
-          <h4 className="product__title h4">Accessories</h4>
-          <span className="product__counter">100 models</span>
-        </article> */}
       </div>
     </section>
   );
