+++ src/pages/CartPage/CartPage.tsx
@@ -1,93 +1,93 @@
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
-import { actions as cartActions } from '../../features/cartSlice';
+// import { useState } from 'react';
+// import { useNavigate } from 'react-router-dom';
+// import { useDispatch } from 'react-redux';
+// import { actions as cartActions } from '../../features/cartSlice';
 
-import { BackButton } from '../../components/BackButton';
-import { CartItem } from '../../components/CartItem';
-import { CheckoutModal } from '../../components/CheckoutModal';
+// import { BackButton } from '../../components/BackButton';
+// import { CartItem } from '../../components/CartItem';
+// import { CheckoutModal } from '../../components/CheckoutModal';
 
-import { useAppSelector } from '../../app/hooks';
+// import { useAppSelector } from '../../app/hooks';
 
-import './CartPage.scss';
-import '../../styles/blocks/button.scss';
-import '../../styles/utils/text-styles.scss';
+// import './CartPage.scss';
+// import '../../styles/blocks/button.scss';
+// import '../../styles/utils/text-styles.scss';
 
-export const CartPage = () => {
-  const dispatch = useDispatch();
-  const products = useAppSelector((state) => state.cartReducer);
+// export const CartPage = () => {
+//   const dispatch = useDispatch();
+//   const products = useAppSelector((state) => state.cartReducer);
 
-  const [showModal, setShowModal] = useState(false);
+//   const [showModal, setShowModal] = useState(false);
 
-  const navigate = useNavigate();
+//   const navigate = useNavigate();
 
-  const totalCost = products.reduce(
-    (total, { product, amount }) => total + product.price * amount,
-    0
-  );
+//   const totalCost = products.reduce(
+//     (total, { product, amount }) => total + product.price * amount,
+//     0
+//   );
 
-  const totalQuantity = products
-    .reduce((total, { amount }) => total + amount, 0);
+//   const totalQuantity = products
+//     .reduce((total, { amount }) => total + amount, 0);
 
-  const handleCheckout = () => {
-    dispatch(cartActions.resetCart());
+//   const handleCheckout = () => {
+//     dispatch(cartActions.resetCart());
 
-    setShowModal(true);
+//     setShowModal(true);
 
-    setTimeout(() => {
-      setShowModal(false);
-      navigate('..');
-    }, 2000);
-  };
+//     setTimeout(() => {
+//       setShowModal(false);
+//       navigate('..');
+//     }, 2000);
+//   };
 
-  return (
-    <div className="cart">
-      <div className="cart__header">
-        <div className="cart__back-button">
-          <BackButton />
-        </div>
+//   return (
+//     <div className="cart">
+//       <div className="cart__header">
+//         <div className="cart__back-button">
+//           <BackButton />
+//         </div>
 
-        <h1 className="cart__title h1">Cart</h1>
-      </div>
+//         <h1 className="cart__title h1">Cart</h1>
+//       </div>
 
-      <div className="cart__items">
-        {products.map(({ product, amount }) => (
-          product ? (
-            <CartItem
-              key={product.id}
-              product={product}
-              quantity={amount}
-            />
-          ) : null
-        ))}
-      </div>
+//       <div className="cart__items">
+//         {products.map(({ product, amount }) => (
+//           product ? (
+//             <CartItem
+//               key={product.id}
+//               product={product}
+//               quantity={amount}
+//             />
+//           ) : null
+//         ))}
+//       </div>
 
-      {!!totalQuantity && (
-        <article className="total">
-          <div className="total__info">
-            <h2 className="total__price">${totalCost}</h2>
-            <div className="total__count-items">
-              Total for
-              {' '}
-              {totalQuantity}
-              {' '}
-              items
-            </div>
-          </div>
+//       {!!totalQuantity && (
+//         <article className="total">
+//           <div className="total__info">
+//             <h2 className="total__price">${totalCost}</h2>
+//             <div className="total__count-items">
+//               Total for
+//               {' '}
+//               {totalQuantity}
+//               {' '}
+//               items
+//             </div>
+//           </div>
 
-          <span className="total__line" />
+//           <span className="total__line" />
 
-          <button
-            type="button"
-            className="button button-add button--cart"
-            onClick={handleCheckout}
-          >
-            Checkout
-          </button>
-        </article>
-      )}
+//           <button
+//             type="button"
+//             className="button button-add button--cart"
+//             onClick={handleCheckout}
+//           >
+//             Checkout
+//           </button>
+//         </article>
+//       )}
 
-      {showModal && <CheckoutModal />}
-    </div>
-  );
-};
+//       {showModal && <CheckoutModal />}
+//     </div>
+//   );
+// };
